devided app.py to different files for easier go through
