+++ extensions.py
@@ -1,0 +1,4 @@
+from flask_socketio import SocketIO
+
+
+socketio = SocketIO(async_mode="threading")
